Add optional incident id to query
