+++ src/sentry/static/sentry/app/components/issues/groupList.tsx
@@ -21,9 +21,10 @@ import {Group} from 'app/types';
 import {callIfFunction} from 'app/utils/callIfFunction';
 import StreamManager from 'app/utils/streamManager';
 import withApi from 'app/utils/withApi';
-import {TimePeriodType} from 'app/views/alerts/rules/details/body';
 
+import {TimePeriodType} from './constants';
 import GroupListHeader from './groupListHeader';
+
 
 const defaultProps = {
   canSelectGroups: true,
