+++ src/sentry/static/sentry/app/views/alerts/rules/details/constants.tsx
@@ -20,3 +20,11 @@ export const TIME_WINDOWS = {
 
 export const API_INTERVAL_POINTS_LIMIT = 10000;
 export const API_INTERVAL_POINTS_MIN = 150;
+
+export type TimePeriodType = {
+  start: string;
+  end: string;
+  period: string;
+  label: string;
+  custom?: boolean;
+};
